chore(license): add Creative Commons Attribution-NonCommercial 4.0 license
Added the Creative Commons Attribution-NonCommercial 4.0 International license to the project. This change allows users to share and adapt the material while ensuring proper attribution and non-commercial use. The inclusion of this license clarifies the terms under which the content can be used, promoting open access while protecting the creator's rights. There are no immediate risks or side effects associated with this addition.