docs(readme): add "go back to table of contents" links for navigation
This update introduces "go back to table of contents" links in each section
of the README file. The addition enhances user navigation, allowing readers
to quickly return to the main table of contents from any section. This
improvement aims to provide a better user experience and facilitate easier
access to information within the document. No risks or side effects are
anticipated from this change.